+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 -   KFP SDK version bumped to 1.8.11 in order to fix misbehaving TTL issue
 -   Dropped support for VertexAI, please use [kedro-vertexi](https://kedro-kubeflow.readthedocs.io/en/latest/index.html) instead
+-   Add Kedro environment name to the pipeline name during upload
 
 ## [0.6.2] - 2022-03-10